requirements.txt
streamlit
python-dotenv
langchain-community
langchain-classic
langchain-core
langchain-openai
langchain-text-splitters
faiss-cpu
pypdf
tiktoken
sentence-transformers
transformers
torch
huggingface-hub
